Create the basic messaging system.
This is just a prototype, but I'm expecting the messaging system to be
an integral part of the design.  This initial implementation should be
compatible with future networking upgrades, because it uses address
numbers instead of pointers.

Under this system, the tokens can only receive.  Game systems can send
and receive messages.