Add the Dot and Tribe classes.
This commit beefs up the token package in preparation of trying to get
the dots to move around.  The Tribe class maintains a list of all the
dots associated with it, and the Dot class knows how to move and how to
draw itself.

The next step is getting the GUI manager to issue the proper commands.