Fix various bugs.
The program now runs in full-screen mode, which makes the controls much
more useable.  I also added the ability to shift-click for waypoints and
to press 'x' to stop the dots.

Handling the shift modifier was a little awkward.  Ultimately, I had to
use a member variable to store the state of the shift key.  I won't want
to do that in the C++ version, though.