Add a combat manager.
I added a combat and/or health manager, so dots can now fight against
each other.  Architecture-wise, I'm starting to wonder if the world
should itself be a manager.  It is already responding to kill/death
messages, and it will be responding to create/birth messages soon.  On
top of that, it has the same update() and draw() methods as the rest of
the managers.

Just food for thought...