Manage exponential dot growth.
Added a growth manager to complement the combat manager.  Once I add
victory conditions, the game will be playable.  In the sense that you
can develop an army of dots and use it to attack your opponent.

The biggest concern revealed by the growth manager has to do with
balance.  How do I keep the number of dots in check, especially with
exponential growth?  Should I have growth rates depend on the total game
population?  That might help force players to fight.

In addition, what should the precise algorithm be for determining the
rate of growth.  Right now, dots grow at a rate of 1 dot every ten
seconds.  This rate is divided by the number of dots within a 50px
radius plus one.

I'd like this function to depend more on the precise distances between
nearby dots.  In addition, I think the growth rate has to be slower.
This will allow players to develop more elaborate strategies and force
players work more on their economy.

Functionality-wise, the game starts to get noticeably laggy once there
are around 50 dots in play.  No doubt this is due to my shitty
algorithms, but it's something to keep in mind.