End game when victory conditions are met.
Add a victory manager to end the game once only one player is left.
The biggest hack here is how the game actually ends.  Right now, I just
set a flag in the victory manager itself and have main check the flag on
every frame.

This is a bad solution because no messages are sent out.  Thus managers
cannot themselves up and this will be hard to adapt to network play.
The reason I didn't send a message is that it's unclear what I would
send a message to.

The game loop is in the main script, and it cannot receive messages.
Perhaps the victory manager could send a message to itself (so other
manager could subscribe to the same message), receive the message, and
then set a flag that would be checked by the main script.

I didn't do this because it won't be necessary until I start doing the
network programming.  Also, whatever flag is set will have to be present
on the client and server sides.  All in all, I'm not ready to make those
decisions yet.

This commit also contains a bunch of files I forgot to commit last time.