Create an AI that moves dots completely randomly.
In order to better playtest the game, it was necessary to create an AI
that would move its dots.  It is no challenge to track down and destroy
stationary dots.  In contrast, it can be very challenging to corner
randomly moving enemies.  This version will make it possible to really
begin playtesting.

One problem with this AI is that it is still very unrealistic.  No human
player could ever control all their dots simultaneously as this AI does.
In addition, a human player would certainly manage dots in groups, which
this AI does not.  Furthermore, random motion tends to cause the AI to
accumulate along the edges of the map, which is also unrealistic.

Hopefully, some of these issuea will be alleviated in the next few days.
This is, however, still just a prototype.